Change the way systems are handled and the main loop starts running.
Systems can now be either generic or class-member functions taking a float and a World reference.

These functions will be provided with the time elapsed and a reference to the current world each tick.

The user is responsible for running the application class with Run() in GameInit. This allows the user to make use of the stack.